moved utility functions to the already existing helpers/functions dir
