+++ apps/skde/pages/helseatlas/kulelinje/index.tsx
@@ -3,9 +3,9 @@ import Skeleton from "@mui/material/Skeleton";
 import BeadLine from "../../../src/components/BeadLine";
 import { FetchMap } from "../../../src/helpers/hooks";
 import { Box } from "@mui/material";
-import getDataUrl from "../utils/getDataUrl";
-import ensureValidLang from "../utils/ensureValidLang";
-import fetchMapPendingOrFailed from "../utils/fetchMapPendingOrFailed";
+import getDataUrl from "../../../src/helpers/functions/getDataUrl";
+import ensureValidLang from "../../../src/helpers/functions/ensureValidLang";
+import fetchMapPendingOrFailed from "../../../src/helpers/functions/fetchMapPendingOrFailed";
 
 const skeleton = (
   <div style={{ display: "flex", justifyContent: "center" }}>
