+++ apps/skde/src/helpers/functions/fetchMapPendingOrFailed.ts
