+++ apps/skde/src/helpers/functions/getDataUrl.ts
